# catalog/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Вывод всех товаров",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Массив со свойствами товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Карточка любого товара",
                "operationId": "5578a4ff4cc0048eefc86673f190bea7",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID товара",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Карточка товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/admin/{category}/{sub_category}/product/create": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Создание товара в разделе подкатегории",
                "operationId": "57b2d5e0802fbed9296588fe7063861e",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID категории",
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "sub_category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID подкатегории",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Карточка товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/admin/products/create": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Создание нового товара",
                "operationId": "a4ccb5a2fdbc6bd366dec29283c474a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Карточка товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductRequest"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/admin/products/{product}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Обновление товара",
                "operationId": "179b156cb88a0b03bf77a7c6e7387e35",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID товара",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Карточка товара",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Удаление товара",
                "operationId": "59266359abd431af75b386e0552b71e6",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID товара",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Товар удален"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/products/find": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Поиск товаров по массиву ID",
                "operationId": "ee2b7624e0824affd9233f7ef4bac1d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": "3"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Массив товаров",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductRequest": {
                "description": "Класс продукта",
                "required": [
                    "name",
                    "description",
                    "price",
                    "category_id",
                    "sub_category_id"
                ],
                "properties": {
                    "name": {
                        "description": "Название продукта",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "description": {
                        "description": "Описание продукта",
                        "type": "text",
                        "example": "Хлопок, сертифицированный по экологическим стандартам, при его производстве не используются агрессивные химикаты. Выбирая такой товар, вы проявляете заботу о себе и о планете."
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "integer",
                        "example": 9990
                    },
                    "previousPrice": {
                        "description": "Прежняя цена продукта",
                        "type": "integer",
                        "example": 10990
                    },
                    "sub_category_id": {
                        "description": "ID подкатегории товара",
                        "type": "integer",
                        "example": 3
                    },
                    "category_id": {
                        "description": "ID категории товара",
                        "type": "integer",
                        "example": 3
                    },
                    "genders": {
                        "description": "Пол товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "gender": {
                                    "type": "string",
                                    "example": "Мужское"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "sizes": {
                        "description": "Размеры товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "size": {
                                    "type": "string",
                                    "example": "XXS"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "colors": {
                        "description": "Цвета товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "color": {
                                    "type": "string",
                                    "example": "#0297c8"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "images": {
                        "description": "Изображения товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "path": {
                                    "type": "string",
                                    "example": "public/images/mockup.png"
                                },
                                "product_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "color_id": {
                                    "type": "integer",
                                    "example": 3
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "ProductResource": {
                "description": "Класс продукта",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "category_id",
                    "sub_category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID продукта",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название продукта",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "description": {
                        "description": "Описание продукта",
                        "type": "text",
                        "example": "Хлопок, сертифицированный по экологическим стандартам, при его производстве не используются агрессивные химикаты. Выбирая такой товар, вы проявляете заботу о себе и о планете."
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "integer",
                        "example": 9990
                    },
                    "previousPrice": {
                        "description": "Прежняя цена продукта",
                        "type": "integer",
                        "example": 10990
                    },
                    "sub_category_id": {
                        "description": "ID подкатегории товара",
                        "type": "integer",
                        "example": 3
                    },
                    "category_id": {
                        "description": "ID категории товара",
                        "type": "integer",
                        "example": 3
                    },
                    "genders": {
                        "description": "Пол товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "gender": {
                                    "type": "string",
                                    "example": "Мужское"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "sizes": {
                        "description": "Размеры товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "size": {
                                    "type": "string",
                                    "example": "XXS"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "colors": {
                        "description": "Цвета товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "color": {
                                    "type": "string",
                                    "example": "#0297c8"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "images": {
                        "description": "Изображения товара",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "path": {
                                    "type": "string",
                                    "example": "public/images/mockup.png"
                                },
                                "product_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "color_id": {
                                    "type": "integer",
                                    "example": 3
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}